+++ src/Common/EngineCore/Systems/Player/Components/Indexers/PlayerComponentEventFilter.cs
@@ -39,6 +39,6 @@ public class PlayerComponentEventFilter<T> : BaseComponentEventFilter<T>
 
     protected override bool ShouldAccept(ulong entityId)
     {
-        return playerTags.HasTagForEntity(entityId);
+        return playerTags.HasTagForEntity(entityId) || playerTags.HasPendingTagForEntity(entityId);
     }
 }